Security authentication program
Security authentication program that allows the user to enter, encrypt, decrypt and compare the authorised access code. The access code is 4 single-digit integer numbers between 0-9. The program allows the user to enter a code, encrypt the number and compare it to an authorised access code as well as allows the user to decrypt an already encrypted code.